+++ libc/string.h
@@ -1,7 +1,7 @@
 #ifndef STRINGS_H
 #define STRINGS_H
 
-void int_to_ascii(int n, char str[]);
+void int_to_ascii(unsigned long n, char str[]);
 void reverse(char s[]);
 int strlen(char s[]);
 void backspace(char s[]);
